requirements.txt
# Core scientific stack
numpy>=1.24,<2.1
pandas>=2.0,<2.3
scipy>=1.10,<1.14
scikit-learn>=1.3,<1.6

# Deep learning
torch>=2.1,<2.5
transformers>=4.35,<4.45
sentence-transformers>=2.2,<3.1

# Retrieval
faiss-cpu>=1.7.4,<1.9
lightgbm>=4.0,<4.6

# Images
pillow>=10.0,<11.0

# Serving
fastapi>=0.104,<0.115
uvicorn[standard]>=0.24,<0.32
python-multipart>=0.0.6

# Caching / utilities
joblib>=1.3,<1.5
pyyaml>=6.0,<7.0
tqdm>=4.66,<5.0

# Visualization
matplotlib>=3.7,<3.10
seaborn>=0.13,<0.14

# Testing
pytest>=7.4,<9.0
pytest-cov>=4.1,<6.0
httpx>=0.25,<0.28

# Dev tooling (kept minimal — pinned soft so CI can install)
ruff>=0.1.6,<0.7
black>=23.10,<25.0
